C Division not populating bug fix
There was some weird logic while traversing rows where if a row had no
sailor, it would skip to the next row. The second row of each school
in C division (this row is also the last one for each school) has no
sailor. The school competitors are put into the dictionary when the
last row is traversed and because of the weird logic, it was being
skipped. Now the logic works.